refactor(resume): motion variant 변수명 수정
