+++ apps/resume/src/components/stack/motion-stack.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Stack } from "./stack";
 
 import { motion } from "framer-motion";
-import { stagger } from "../../constants/motion.config";
+import { staggerContainer } from "../../constants/motion-variants";
 
 interface MotionStackProps {
   id: string;
@@ -14,7 +14,7 @@ export function MotionStack({ id, title, children }: MotionStackProps) {
   return (
     <Stack id={id} title={title}>
       <motion.div
-        variants={stagger}
+        variants={staggerContainer}
         initial="hidden"
         animate="show"
         className="grid gap-3"
